All updated codes related to Network Graphs
